Revert "0.10.2 Patch Release (#453)"
This reverts commit 77c9175647c22caa1413cc9f507d9fc9ad4c1f69.
