+++ aws-opentelemetry-distro/src/amazon/opentelemetry/distro/version.py
@@ -1,4 +1,4 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-__version__ = "0.10.1.dev0"
+__version__ = "0.10.1"
