+++ aws-opentelemetry-distro/tests/amazon/opentelemetry/distro/test_aws_opentelementry_configurator.py
@@ -25,7 +25,6 @@ from amazon.opentelemetry.distro.aws_opentelemetry_configurator import (
     AwsOpenTelemetryConfigurator,
     OtlpLogHeaderSetting,
     _check_emf_exporter_enabled,
-    _clear_logs_header_cache,
     _create_aws_otlp_exporter,
     _create_emf_exporter,
     _custom_import_sampler,
@@ -38,16 +37,16 @@ from amazon.opentelemetry.distro.aws_opentelemetry_configurator import (
     _customize_span_processors,
     _export_unsampled_span_for_agent_observability,
     _export_unsampled_span_for_lambda,
-    _fetch_logs_header,
     _init_logging,
     _is_application_signals_enabled,
     _is_application_signals_runtime_enabled,
     _is_defer_to_workers_enabled,
     _is_wsgi_master_process,
+    _validate_and_fetch_logs_header,
 )
 from amazon.opentelemetry.distro.aws_opentelemetry_distro import AwsOpenTelemetryDistro
 from amazon.opentelemetry.distro.aws_span_metrics_processor import AwsSpanMetricsProcessor
-from amazon.opentelemetry.distro.exporter.console.logs.compact_console_log_exporter import CompactConsoleLogExporter
+from amazon.opentelemetry.distro.exporter.aws.metrics.aws_cloudwatch_emf_exporter import AwsCloudWatchEmfExporter
 from amazon.opentelemetry.distro.exporter.otlp.aws.common.aws_auth_session import AwsAuthSession
 
 # pylint: disable=line-too-long
@@ -71,7 +70,7 @@ from opentelemetry.exporter.otlp.proto.http.metric_exporter import OTLPMetricExp
 from opentelemetry.exporter.otlp.proto.http.trace_exporter import OTLPSpanExporter
 from opentelemetry.metrics import get_meter_provider
 from opentelemetry.processor.baggage import BaggageSpanProcessor
-from opentelemetry.sdk._logs.export import BatchLogRecordProcessor, ConsoleLogExporter
+from opentelemetry.sdk._logs.export import BatchLogRecordProcessor
 from opentelemetry.sdk.environment_variables import OTEL_TRACES_SAMPLER, OTEL_TRACES_SAMPLER_ARG
 from opentelemetry.sdk.metrics._internal.export import PeriodicExportingMetricReader
 from opentelemetry.sdk.resources import Resource
@@ -509,7 +508,6 @@ class TestAwsOpenTelemetryConfigurator(TestCase):
             bad_configs.append(config)
 
         for config in good_configs:
-            _clear_logs_header_cache()
             self.customize_exporter_test(
                 config,
                 _customize_span_exporter,
@@ -521,7 +519,6 @@ class TestAwsOpenTelemetryConfigurator(TestCase):
             )
 
         for config in bad_configs:
-            _clear_logs_header_cache()
             self.customize_exporter_test(
                 config,
                 _customize_span_exporter,
@@ -619,7 +616,6 @@ class TestAwsOpenTelemetryConfigurator(TestCase):
             bad_configs.append(config)
 
         for config in good_configs:
-            _clear_logs_header_cache()
             self.customize_exporter_test(
                 config,
                 _customize_logs_exporter,
@@ -630,7 +626,6 @@ class TestAwsOpenTelemetryConfigurator(TestCase):
             )
 
         for config in bad_configs:
-            _clear_logs_header_cache()
             self.customize_exporter_test(
                 config, _customize_logs_exporter, OTLPLogExporter(), OTLPLogExporter, Session, Compression.NoCompression
             )
@@ -680,58 +675,6 @@ class TestAwsOpenTelemetryConfigurator(TestCase):
             self.assertIsInstance(captured_exporter, expected_exporter)
 
         os.environ.pop(OTEL_EXPORTER_OTLP_LOGS_ENDPOINT)
-
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator.LoggingHandler", return_value=MagicMock())
-    @patch("logging.getLogger", return_value=MagicMock())
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._customize_logs_exporter")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator.LoggerProvider", return_value=MagicMock())
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._customize_log_record_processor")
-    def test_init_logging_console_exporter_replacement(
-        self,
-        mock_customize_processor,
-        mock_logger_provider,
-        mock_customize_logs_exporter,
-        mock_get_logger,
-        mock_logging_handler,
-    ):
-        """Test that ConsoleLogExporter is replaced with CompactConsoleLogExporter when in Lambda"""
-
-        # Mock _is_lambda_environment to return True
-        with patch(
-            "amazon.opentelemetry.distro.aws_opentelemetry_configurator._is_lambda_environment", return_value=True
-        ):
-            # Test with ConsoleLogExporter
-            exporters = {"console": ConsoleLogExporter}
-            _init_logging(exporters, Resource.get_empty())
-
-            # Verify that _customize_log_record_processor was called
-            mock_customize_processor.assert_called_once()
-
-            # Get the exporter that was passed to _customize_logs_exporter
-            call_args = mock_customize_logs_exporter.call_args
-            exporter_instance = call_args[0][0]
-
-            # Verify it's a CompactConsoleLogExporter instance
-            self.assertIsInstance(exporter_instance, CompactConsoleLogExporter)
-
-        # Reset mocks
-        mock_customize_processor.reset_mock()
-        mock_customize_logs_exporter.reset_mock()
-
-        # Test when not in Lambda environment - should not replace
-        with patch(
-            "amazon.opentelemetry.distro.aws_opentelemetry_configurator._is_lambda_environment", return_value=False
-        ):
-            exporters = {"console": ConsoleLogExporter}
-            _init_logging(exporters, Resource.get_empty())
-
-            # Get the exporter that was passed to _customize_logs_exporter
-            call_args = mock_customize_logs_exporter.call_args
-            exporter_instance = call_args[0][0]
-
-            # Verify it's still a regular ConsoleLogExporter
-            self.assertIsInstance(exporter_instance, ConsoleLogExporter)
-            self.assertNotIsInstance(exporter_instance, CompactConsoleLogExporter)
 
     def test_customize_span_processors(self):
         mock_tracer_provider: TracerProvider = MagicMock()
@@ -1021,72 +964,64 @@ class TestAwsOpenTelemetryConfigurator(TestCase):
         # Clean up
         os.environ.pop("OTEL_METRICS_EXPORTER", None)
 
-    def test_fetch_logs_header(self):
-        _clear_logs_header_cache()
-
+    def test_validate_and_fetch_logs_header(self):
         # Test when headers are not set
         os.environ.pop(OTEL_EXPORTER_OTLP_LOGS_HEADERS, None)
-        result = _fetch_logs_header()
+        result = _validate_and_fetch_logs_header()
         self.assertIsInstance(result, OtlpLogHeaderSetting)
         self.assertIsNone(result.log_group)
         self.assertIsNone(result.log_stream)
         self.assertIsNone(result.namespace)
-        self.assertFalse(result.is_valid())
+        self.assertFalse(result.is_valid)
 
-        # Test singleton behavior - should return the same cached instance
-        result2 = _fetch_logs_header()
-        self.assertIs(result, result2)  # Same object reference
-
-        _clear_logs_header_cache()
+        # Test with valid headers
         os.environ[OTEL_EXPORTER_OTLP_LOGS_HEADERS] = "x-aws-log-group=test-group,x-aws-log-stream=test-stream"
-        result = _fetch_logs_header()
+        result = _validate_and_fetch_logs_header()
         self.assertEqual(result.log_group, "test-group")
         self.assertEqual(result.log_stream, "test-stream")
         self.assertIsNone(result.namespace)
-        self.assertTrue(result.is_valid())
+        self.assertTrue(result.is_valid)
 
-        # Test singleton behavior again
-        result2 = _fetch_logs_header()
-        self.assertIs(result, result2)
-
-        _clear_logs_header_cache()
+        # Test with valid headers including namespace
         os.environ[OTEL_EXPORTER_OTLP_LOGS_HEADERS] = (
             "x-aws-log-group=test-group,x-aws-log-stream=test-stream,x-aws-metric-namespace=test-namespace"
         )
-        result = _fetch_logs_header()
-        self.assertEqual(result.namespace, "test-namespace")
-        self.assertTrue(result.is_valid())
-
-        _clear_logs_header_cache()
-        os.environ[OTEL_EXPORTER_OTLP_LOGS_HEADERS] = "x-aws-log-stream=test-stream"
-        result = _fetch_logs_header()
-        self.assertEqual(result.log_stream, "test-stream")
-        self.assertFalse(result.is_valid())
-
-        _clear_logs_header_cache()
-        os.environ[OTEL_EXPORTER_OTLP_LOGS_HEADERS] = "x-aws-log-group=test-group"
-        result = _fetch_logs_header()
+        result = _validate_and_fetch_logs_header()
         self.assertEqual(result.log_group, "test-group")
-        self.assertIsNone(result.log_stream)
-        self.assertFalse(result.is_valid())
+        self.assertEqual(result.log_stream, "test-stream")
+        self.assertEqual(result.namespace, "test-namespace")
+        self.assertTrue(result.is_valid)
 
-        _clear_logs_header_cache()
-        os.environ[OTEL_EXPORTER_OTLP_LOGS_HEADERS] = "x-aws-log-group=,x-aws-log-stream=test-stream"
-        result = _fetch_logs_header()
+        # Test with missing log group
+        os.environ[OTEL_EXPORTER_OTLP_LOGS_HEADERS] = "x-aws-log-stream=test-stream"
+        result = _validate_and_fetch_logs_header()
         self.assertIsNone(result.log_group)
         self.assertEqual(result.log_stream, "test-stream")
-        self.assertFalse(result.is_valid())
+        self.assertFalse(result.is_valid)
 
-        _clear_logs_header_cache()
-        os.environ[OTEL_EXPORTER_OTLP_LOGS_HEADERS] = "x-aws-log-group=test-group,x-aws-log-stream="
-        result = _fetch_logs_header()
+        # Test with missing log stream
+        os.environ[OTEL_EXPORTER_OTLP_LOGS_HEADERS] = "x-aws-log-group=test-group"
+        result = _validate_and_fetch_logs_header()
         self.assertEqual(result.log_group, "test-group")
         self.assertIsNone(result.log_stream)
-        self.assertFalse(result.is_valid())
+        self.assertFalse(result.is_valid)
+
+        # Test with empty value in log group
+        os.environ[OTEL_EXPORTER_OTLP_LOGS_HEADERS] = "x-aws-log-group=,x-aws-log-stream=test-stream"
+        result = _validate_and_fetch_logs_header()
+        self.assertIsNone(result.log_group)
+        self.assertEqual(result.log_stream, "test-stream")
+        self.assertFalse(result.is_valid)
+
+        # Test with empty value in log stream
+        os.environ[OTEL_EXPORTER_OTLP_LOGS_HEADERS] = "x-aws-log-group=test-group,x-aws-log-stream="
+        result = _validate_and_fetch_logs_header()
+        self.assertEqual(result.log_group, "test-group")
+        self.assertIsNone(result.log_stream)
+        self.assertFalse(result.is_valid)
 
         # Clean up
         os.environ.pop(OTEL_EXPORTER_OTLP_LOGS_HEADERS, None)
-        _clear_logs_header_cache()
 
     @patch(
         "amazon.opentelemetry.distro.aws_opentelemetry_configurator.is_agent_observability_enabled", return_value=False
@@ -1115,6 +1050,63 @@ class TestAwsOpenTelemetryConfigurator(TestCase):
         mock_logger_provider.add_log_record_processor.assert_called_once()
         added_processor = mock_logger_provider.add_log_record_processor.call_args[0][0]
         self.assertIsInstance(added_processor, AwsCloudWatchOtlpBatchLogRecordProcessor)
+
+    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._validate_and_fetch_logs_header")
+    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator.get_aws_session")
+    def test_create_emf_exporter(self, mock_get_session, mock_validate):
+        # Test when botocore is not installed
+        mock_get_session.return_value = None
+        result = _create_emf_exporter()
+        self.assertIsNone(result)
+
+        # Reset mock for subsequent tests
+        mock_get_session.reset_mock()
+        mock_get_session.return_value = MagicMock()
+
+        # Mock the EMF exporter class import by patching the module import
+        with patch(
+            "amazon.opentelemetry.distro.exporter.aws.metrics.aws_cloudwatch_emf_exporter.AwsCloudWatchEmfExporter"
+        ) as mock_emf_exporter_class:
+            mock_exporter_instance = MagicMock()
+            mock_exporter_instance.namespace = "default"
+            mock_exporter_instance.log_group_name = "test-group"
+            mock_emf_exporter_class.return_value = mock_exporter_instance
+
+            # Test when headers are invalid
+            mock_validate.return_value = OtlpLogHeaderSetting(None, None, None, False)
+            result = _create_emf_exporter()
+            self.assertIsNone(result)
+
+            # Test when namespace is missing (should still create exporter with default namespace)
+            mock_validate.return_value = OtlpLogHeaderSetting("test-group", "test-stream", None, True)
+            result = _create_emf_exporter()
+            self.assertIsNotNone(result)
+            self.assertEqual(result, mock_exporter_instance)
+            # Verify that the EMF exporter was called with correct parameters
+            mock_emf_exporter_class.assert_called_with(
+                session=mock_get_session.return_value,
+                namespace=None,
+                log_group_name="test-group",
+                log_stream_name="test-stream",
+            )
+
+            # Test with valid configuration
+            mock_validate.return_value = OtlpLogHeaderSetting("test-group", "test-stream", "test-namespace", True)
+            result = _create_emf_exporter()
+            self.assertIsNotNone(result)
+            self.assertEqual(result, mock_exporter_instance)
+            # Verify that the EMF exporter was called with correct parameters
+            mock_emf_exporter_class.assert_called_with(
+                session=mock_get_session.return_value,
+                namespace="test-namespace",
+                log_group_name="test-group",
+                log_stream_name="test-stream",
+            )
+
+            # Test exception handling
+            mock_validate.side_effect = Exception("Test exception")
+            result = _create_emf_exporter()
+            self.assertIsNone(result)
 
     @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator.get_logger_provider")
     @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator.is_agent_observability_enabled")
@@ -1182,6 +1174,35 @@ class TestAwsOpenTelemetryConfigurator(TestCase):
         result = _create_aws_otlp_exporter("https://xray.us-east-1.amazonaws.com/v1/traces", "xray", "us-east-1")
         self.assertIsNone(result)
 
+    def test_customize_metric_exporters_with_emf(self):
+        metric_readers = []
+        views = []
+
+        # Test with EMF disabled
+        _customize_metric_exporters(metric_readers, views, is_emf_enabled=False)
+        self.assertEqual(len(metric_readers), 0)
+
+        # Test with EMF enabled but create_emf_exporter returns None
+        with patch(
+            "amazon.opentelemetry.distro.aws_opentelemetry_configurator._create_emf_exporter", return_value=None
+        ):
+            _customize_metric_exporters(metric_readers, views, is_emf_enabled=True)
+            self.assertEqual(len(metric_readers), 0)
+
+        # Test with EMF enabled and valid exporter
+        mock_emf_exporter = MagicMock(spec=AwsCloudWatchEmfExporter)
+        # Add the required attributes that PeriodicExportingMetricReader expects
+        mock_emf_exporter._preferred_temporality = {}
+        mock_emf_exporter._preferred_aggregation = {}
+
+        with patch(
+            "amazon.opentelemetry.distro.aws_opentelemetry_configurator._create_emf_exporter",
+            return_value=mock_emf_exporter,
+        ):
+            _customize_metric_exporters(metric_readers, views, is_emf_enabled=True)
+            self.assertEqual(len(metric_readers), 1)
+            self.assertIsInstance(metric_readers[0], PeriodicExportingMetricReader)
+
     @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator.is_agent_observability_enabled")
     @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator.get_service_attribute")
     def test_customize_resource_without_agent_observability(self, mock_get_service_attribute, mock_is_agent_enabled):
@@ -1228,202 +1249,6 @@ class TestAwsOpenTelemetryConfigurator(TestCase):
         # Should preserve existing agent type and not override it
         self.assertEqual(result.attributes[AWS_LOCAL_SERVICE], "test-service")
         self.assertEqual(result.attributes[AWS_SERVICE_TYPE], "existing-agent")
-
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._fetch_logs_header")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._is_lambda_environment")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator.get_aws_session")
-    def test_create_emf_exporter_lambda_without_valid_headers(
-        self, mock_get_session, mock_is_lambda, mock_fetch_headers
-    ):
-        """Test _create_emf_exporter returns ConsoleEmfExporter for Lambda without valid log headers"""
-        # Setup mocks
-        mock_is_lambda.return_value = True
-        mock_header_setting = MagicMock()
-        mock_header_setting.is_valid.return_value = False
-        mock_header_setting.namespace = "test-namespace"
-        mock_fetch_headers.return_value = mock_header_setting
-
-        with patch(
-            "amazon.opentelemetry.distro.exporter.aws.metrics.console_emf_exporter.ConsoleEmfExporter"
-        ) as mock_console_exporter:
-            mock_exporter_instance = MagicMock()
-            mock_console_exporter.return_value = mock_exporter_instance
-
-            result = _create_emf_exporter()
-
-            self.assertEqual(result, mock_exporter_instance)
-            mock_console_exporter.assert_called_once_with(namespace="test-namespace")
-
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._fetch_logs_header")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._is_lambda_environment")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator.get_aws_session")
-    def test_create_emf_exporter_lambda_with_valid_headers(self, mock_get_session, mock_is_lambda, mock_fetch_headers):
-        """Test _create_emf_exporter returns AwsCloudWatchEmfExporter for Lambda with valid headers"""
-        # Setup mocks
-        mock_is_lambda.return_value = True
-        mock_session = MagicMock()
-        mock_get_session.return_value = mock_session
-
-        mock_header_setting = MagicMock()
-        mock_header_setting.is_valid.return_value = True
-        mock_header_setting.namespace = "test-namespace"
-        mock_header_setting.log_group = "test-group"
-        mock_header_setting.log_stream = "test-stream"
-        mock_fetch_headers.return_value = mock_header_setting
-
-        with patch(
-            "amazon.opentelemetry.distro.exporter.aws.metrics.aws_cloudwatch_emf_exporter.AwsCloudWatchEmfExporter"
-        ) as mock_cloudwatch_exporter:
-            mock_exporter_instance = MagicMock()
-            mock_cloudwatch_exporter.return_value = mock_exporter_instance
-
-            result = _create_emf_exporter()
-
-            self.assertEqual(result, mock_exporter_instance)
-            mock_cloudwatch_exporter.assert_called_once_with(
-                session=mock_session,
-                namespace="test-namespace",
-                log_group_name="test-group",
-                log_stream_name="test-stream",
-            )
-
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._fetch_logs_header")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._is_lambda_environment")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator.get_aws_session")
-    def test_create_emf_exporter_non_lambda_with_valid_headers(
-        self, mock_get_session, mock_is_lambda, mock_fetch_headers
-    ):
-        """Test _create_emf_exporter returns AwsCloudWatchEmfExporter for non-Lambda with valid headers"""
-        # Setup mocks
-        mock_is_lambda.return_value = False
-        mock_session = MagicMock()
-        mock_get_session.return_value = mock_session
-
-        mock_header_setting = MagicMock()
-        mock_header_setting.is_valid.return_value = True
-        mock_header_setting.namespace = "test-namespace"
-        mock_header_setting.log_group = "test-group"
-        mock_header_setting.log_stream = "test-stream"
-        mock_fetch_headers.return_value = mock_header_setting
-
-        with patch(
-            "amazon.opentelemetry.distro.exporter.aws.metrics.aws_cloudwatch_emf_exporter.AwsCloudWatchEmfExporter"
-        ) as mock_cloudwatch_exporter:
-            mock_exporter_instance = MagicMock()
-            mock_cloudwatch_exporter.return_value = mock_exporter_instance
-
-            result = _create_emf_exporter()
-
-            self.assertEqual(result, mock_exporter_instance)
-            mock_cloudwatch_exporter.assert_called_once_with(
-                session=mock_session,
-                namespace="test-namespace",
-                log_group_name="test-group",
-                log_stream_name="test-stream",
-            )
-
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._fetch_logs_header")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._is_lambda_environment")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator.get_aws_session")
-    def test_create_emf_exporter_non_lambda_without_valid_headers(
-        self, mock_get_session, mock_is_lambda, mock_fetch_headers
-    ):
-        """Test _create_emf_exporter returns None for non-Lambda without valid headers"""
-        # Setup mocks
-        mock_is_lambda.return_value = False
-        mock_session = MagicMock()
-        mock_get_session.return_value = mock_session
-
-        mock_header_setting = MagicMock()
-        mock_header_setting.is_valid.return_value = False
-        mock_fetch_headers.return_value = mock_header_setting
-
-        result = _create_emf_exporter()
-
-        self.assertIsNone(result)
-
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._fetch_logs_header")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._is_lambda_environment")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator.get_aws_session")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._logger")
-    def test_create_emf_exporter_no_botocore_session(
-        self, mock_logger, mock_get_session, mock_is_lambda, mock_fetch_headers
-    ):
-        """Test _create_emf_exporter returns None when botocore session is not available"""
-        # Setup mocks
-        mock_is_lambda.return_value = False
-        mock_get_session.return_value = None  # Simulate missing botocore
-
-        mock_header_setting = MagicMock()
-        mock_header_setting.is_valid.return_value = True
-        mock_fetch_headers.return_value = mock_header_setting
-
-        result = _create_emf_exporter()
-
-        self.assertIsNone(result)
-        mock_logger.warning.assert_called_once_with("botocore is not installed. EMF exporter requires botocore")
-
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._fetch_logs_header")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._logger")
-    def test_create_emf_exporter_exception_handling(self, mock_logger, mock_fetch_headers):
-        """Test _create_emf_exporter handles exceptions gracefully"""
-        # Setup mocks to raise exception
-        test_exception = Exception("Test exception")
-        mock_fetch_headers.side_effect = test_exception
-
-        result = _create_emf_exporter()
-
-        self.assertIsNone(result)
-        mock_logger.error.assert_called_once_with("Failed to create EMF exporter: %s", test_exception)
-
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._fetch_logs_header")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._is_lambda_environment")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator.get_aws_session")
-    def test_create_emf_exporter_lambda_without_valid_headers_none_namespace(
-        self, mock_get_session, mock_is_lambda, mock_fetch_headers
-    ):
-        """Test _create_emf_exporter with Lambda environment and None namespace"""
-        # Setup mocks
-        mock_is_lambda.return_value = True
-        mock_header_setting = MagicMock()
-        mock_header_setting.is_valid.return_value = False
-        mock_header_setting.namespace = None
-        mock_fetch_headers.return_value = mock_header_setting
-
-        with patch(
-            "amazon.opentelemetry.distro.exporter.aws.metrics.console_emf_exporter.ConsoleEmfExporter"
-        ) as mock_console_exporter:
-            mock_exporter_instance = MagicMock()
-            mock_console_exporter.return_value = mock_exporter_instance
-
-            result = _create_emf_exporter()
-
-            self.assertEqual(result, mock_exporter_instance)
-            mock_console_exporter.assert_called_once_with(namespace=None)
-
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._fetch_logs_header")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._is_lambda_environment")
-    @patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator.get_aws_session")
-    def test_create_emf_exporter_cloudwatch_exporter_import_error(
-        self, mock_get_session, mock_is_lambda, mock_fetch_headers
-    ):
-        """Test _create_emf_exporter handles import errors for CloudWatch exporter"""
-        # Setup mocks
-        mock_is_lambda.return_value = False
-        mock_session = MagicMock()
-        mock_get_session.return_value = mock_session
-
-        mock_header_setting = MagicMock()
-        mock_header_setting.is_valid.return_value = True
-        mock_fetch_headers.return_value = mock_header_setting
-
-        # Mock import to raise ImportError
-        with patch("amazon.opentelemetry.distro.aws_opentelemetry_configurator._logger") as mock_logger:
-            with patch("builtins.__import__", side_effect=ImportError("Cannot import CloudWatch exporter")):
-                result = _create_emf_exporter()
-
-                self.assertIsNone(result)
-                mock_logger.error.assert_called_once()
 
 
 def validate_distro_environ():
